reorganize app dir, add more initial tests, update README
