+++ README.md
@@ -1,3 +1,46 @@
 # wave-player
 
-An audio player.
+A music player.
+
+## Setup
+
+Install dependencies:
+
+```zsh
+npm i
+# or
+pnpm i
+```
+
+Run development server:
+
+```zsh
+npm run dev
+# or
+pnpm dev
+```
+
+Run tests:
+
+```zsh
+npm run test
+npm run test:watch # watch mode
+# or
+pnpm test
+pnpm test:watch
+```
+
+Build:
+
+```zsh
+npm run build
+# or
+pnpm build
+```
+
+---
+
+## TODOs
+
+- Add more tests
+- Add `WavePlayerStack` context
