+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 import { WavePlayerMode } from '@/lib/types'
-import WavePlayer from './(wave-player)/WavePlayer'
-import WavePlayerStack from './(wave-player)/WavePlayerStack'
+import WavePlayer from './wave-player/WavePlayer'
+import WavePlayerStack from './wave-player/WavePlayerStack'
 import { tracks } from '@/lib/tracks'
 
 export default function Home() {
@@ -17,7 +17,9 @@ export default function Home() {
         />
         <h2 className='mt-6 text-xl font-mono'>WavePlayerStack</h2>
         <WavePlayerStack
+          id={0}
           tracks={tracks}
+          isActive={false}
         />
       </div>
     </main>
