+++ src/app/wave-player/components/ProgressBar.tsx
