+++ src/app/wave-player/components/TrackImage.tsx
