+++ src/app/wave-player/components/TrackInfo.tsx
